EntityBase upgrade
A lot of changes done on EntityBase.cs first tests were made in Play Mode.

Set up scene
Tested the basic mechanics (some working, some not yet)
There are a couple of issues, which I plan to solve soon.

There's still a lot to think over. But a start has been made - although a lot will have to be refactored.